+++ src/assets/blocks/coming-soon/index.asset.php
@@ -1,0 +1,9 @@
+<?php declare( strict_types = 1 );
+/**
+ * Plugin Name: Lumiere Movies Into the post options
+ * Description: Select boxes that will display movies/persons
+ * Version: 1.0
+ *
+ * @package lumieremovies
+ */
+return [ 'dependencies' => [ 'wp-blocks', 'wp-block-editor', 'wp-element', 'wp-components', 'wp-data', 'wp-i18n' ] ];